+++ _doc/sphinxdoc/source/conf.py
@@ -51,6 +51,7 @@ epkg_dictionary['mod_wsgi'] = 'https://github.com/GrahamDumpleton/mod_wsgi'
 epkg_dictionary['PIL'] = ("https://pillow.readthedocs.io/en/4.3.x/",
                           ("https://pillow.readthedocs.io/en/4.3.x/reference/{0}.html", 1),
                           ("https://pillow.readthedocs.io/en/4.3.x/reference/{0}.html#PIL.{0}.{1}", 2))
+epkg_dictionary['pyjwt'] = 'https://pyjwt.readthedocs.io/en/latest/index.html'
 epkg_dictionary['REST API'] = "https://en.wikipedia.org/wiki/Representational_state_transfer"
 epkg_dictionary['uwsgi'] = 'http://uwsgi-docs.readthedocs.io/en/latest/'
 epkg_dictionary['waitress'] = 'https://docs.pylonsproject.org/projects/waitress/en/latest/'