+++ src/lightmlrestapi/mlapp/__init__.py
@@ -4,3 +4,4 @@
 """
 
 from .mlpost import MachineLearningPost
+from .base_logging import enumerate_parsed_logs