+++ HISTORY.rst
@@ -1,12 +1,17 @@
 
+
+.. _l-HISTORY:
+
 =======
 History
 =======
 
-current - 2018-04-14 - 0.00Mb
+current - 2018-04-15 - 0.00Mb
 =============================
 
-* `1`: fix gallery of examples (style) (2018-01-05)
+* `2`: add encrypted logging (2018-04-15) 
+* `1`: fix gallery of examples (style) (2018-01-05) 
 
 0.1.37 - 2017-12-06 - 0.10Mb
 ============================
+